clear tables during migration (see comment)
Foreign key constraints are not active during migration.
At least   db.execSQL("PRAGMA foreign_keys=ON;")  etc did not help.
Because of this it is not enough to just clear the Conversations table (to have cascade deletion in
other tables), but all related tables have to be cleared with SQL statement as well.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>