Add interface for common fields of "overall" websocket messages
The interface is now passed to "sendMessage" instead of a serialized
message, which will make possible to customize the "id" property (still
to be added) if needed before sending the message.

Note that, for consistency, the interface was applied to all "overall"
data objects, even if some of them are only for received messages (like
"ErrorOverallWebSocketMessage").

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>