Add support for running callbacks on signaling message responses
With the external signaling server, if the "id" property is set when
sending a signaling message the response to that specific message will
be identified by the same id. Now a callback can be passed to
"sendMessage", which automatically adds an id to the message and makes
possible to run the callback when/if the response to the message is
received.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>