Fix handling of signaling message responses when joining a room
When joining a room, the signaling server confirms that joining the room
was successful by sending a "room" message. Until now the Android app
assumed that "room" messages were sent only in that case, but a "room"
message can be sent also if the properties of the room change. When the
room is joined while the call activity is active the call is also
joined, so if the properties of the room changed while in a call the
Android app rejoined the call, which caused strange issues due to how it
was done. For example, when a recording started additional guests
appeared in the UI.

To solve that now only the "room" message that actually confirms that
the join was successful, which can be identified by setting an ID in the
request to join the room, is treated as such.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>