remove thread for getTempMessagesForConversation
i'm not yet sure about it and did not have a problem with it, but i guess this change makes sense to ensure temporary messages from a thread are not shown and remain in the main chat. This could theoretically happen when directly leaving a thread after sending a message?

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>