fix to always open correct chat (hopefully)
The flag FLAG_UPDATE_CURRENT should ensure that if a matching PendingIntent already exists, the old one with stale extras is not returned but instead it's updated with the new extras

Because this was missing, it could happen that a wrong chat was opened when clicking on a notification.

I was never able to reproduce it, but the missing flag should explain why there was the bug for some users.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>