set version to 21.2.0
For 21.2.0, loginv2 and qr code login were reverted because loginv2 did not work reliable and qr code login was build on top of the changed code.
To mot mess around, both features are reverted and v21.2.0 will basically only bump the targetSdkVersion and include minor bugfixes.

qr code login will comme with v22 and loginv2 whenever it is ready to be released.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>