modify backpress handling for threads
only when coming from notification and backpress is done in a thread, then open the conversationlist

Introducing the extra openedViaNotification for this is not a nice solution. But for release of 22.0.0 greater changes should be avoided.

Best might be for take care of the backstack completely by activities intent flags instead to override handleOnBackPressed

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>