fix loading spinner for participants without peerConnection
This happens e.g. when they don't have permissions to send audio and video.

The fix is to set isConnected = true in the ParticipantUiState creation.

At first it felt like setting isConnected to true is a hack, but it should be okay as it's immediately changed by the peerConnection handling for the other cases.

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>