WIP analyze why emoji is not sent as styled emoji
There is something wrong with sending the Emoji Variation Selector.

I debugged and saw that there are differences between web and android:

from web:

emoji ❤️
code U+2764
code U+FE0F

from Android:

emoji ❤
code U+2764

So this is about Emoji Variation Selector (VS-16).
The emojis that are send from android are not styled (they do not include the Emoji Variation Selector).

If i append FE0F to the classic emojis, it is fixed. But it seems the challenge is to decide when to append the Emoji Variation Selector.

i quickly tried to append the style if there is not already one while sending. Works for classic emojis to append the style but other styled emojis are duplicated...

pausing for now due to other important things..

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>