WIP mark chat as read
wip:
Maybe trigger it when the user scrolled to the bottom on his own.
E.g. when a scrolling listener tells "scrolling is finished" we check if it's scrolled to the bottom and then send the mark as read

also todo:
mark as read "when the user is sending the app to the background or navigates "back to conversation list" and the last message a user read (saw on screen) is a higher id then the previous known readmarker"

Signed-off-by: Marcel Hibbe <dev@mhibbe.de>